move into own component
