+++ moped-editor/src/views/projects/projectView/ProjectSummary/ProjectSummaryCouncilDistricts.js
@@ -1,0 +1,37 @@
+import React from "react";
+import { Box, Grid, Typography } from "@mui/material";
+
+/**
+ * ProjectSummaryCouncilDistricts Component
+ * @param {Object} data - The project geography from the graphql query's data object
+ * @param {Object} classes - The shared style settings
+ * @returns {JSX.Element}
+ * @constructor
+ */
+const ProjectSummaryCouncilDistricts = ({ projectGeography, classes }) => {
+  // reduce the array of geography objects into an array of city council districts
+  const getDistricts = (data) => {
+    const initialValue = [];
+    const districts = data.reduce(
+      (acc, component) => [...acc, component["council_districts"]],
+      initialValue
+    );
+
+    return [...new Set(districts.flat())];
+  };
+
+  return (
+    <Grid item xs={12} className={classes.fieldGridItem}>
+      <Typography className={classes.fieldLabel}>
+        Council district(s):
+      </Typography>
+      <Box className={classes.fieldBox}>
+        <Typography className={classes.fieldLabelTextNoHover}>
+          {getDistricts(projectGeography).join(", ")}
+        </Typography>
+      </Box>
+    </Grid>
+  );
+};
+
+export default ProjectSummaryCouncilDistricts;
